Update 2018-01-09-SCS.md
Here is the clarification you asked me. I tried to stick close to the original texts, only modifying what I think are important clarifications. 
I also added some link here and there for the things I thought about (as a reader, it's always painful when you have to google things that are mentioned ;)

I am very grateful for this invitation to edit :)